Extract publish function into separate file
